refactor(template): extract DefaultTypePreview to Component
